+++ packages/deployments/contracts/contracts/nomad-xapps/contracts/bridge/BridgeRouter.sol
@@ -16,8 +16,6 @@ import {TypedMemView} from "../../../nomad-core/libs/TypedMemView.sol";
 import {IERC20Upgradeable} from "@openzeppelin/contracts-upgradeable/token/ERC20/IERC20Upgradeable.sol";
 import {SafeERC20Upgradeable} from "@openzeppelin/contracts-upgradeable/token/ERC20/utils/SafeERC20Upgradeable.sol";
 
-import "hardhat/console.sol";
-
 /**
  * @title BridgeRouter
  */
