+++ packages/deployments/contracts/src/tasks/getChainId.ts
@@ -9,12 +9,12 @@ export default task("get-chain-id", "Get chainId")
 
     let connextAddress = _connextAddress;
     if (!connextAddress) {
-      const connextDeployment = await deployments.get("ConnextUpgradeBeaconProxy");
+      const connextDeployment = await deployments.get("Connext_Proxy");
       connextAddress = connextDeployment.address;
     }
     console.log("connextAddress: ", connextAddress);
 
-    const connext = await ethers.getContractAt("Connext", connextAddress);
+    const connext = await ethers.getContractAt("Connext_Implementation", connextAddress);
     const chainId = await connext.getChainId();
     console.log("chainId: ", chainId.toString());
   });
